requirements.txt
numpy
matplotlib
jupyterlab
ruamel.yaml
tqdm
wandb
imageio
seaborn
pygame

# Install also (see `install.md`):
#   1. Our fork of the htm.core
#   2. [Optional] CoppeliaSim + PyRep
#   3. [Optional] AnimalAI
